+++ packages/vercel-sandbox/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @vercel/sandbox
 
+## 2.0.0-beta.13
+
+### Minor Changes
+
+- Support new onResume parameter in Sandbox.create and Sandbox.get
+
+## 2.0.0-beta.12
+
+### Minor Changes
+
+- Rebase from main
+
 ## 2.0.0-beta.11
 
 ### Patch Changes